# docs/swagger.yaml
basePath: /
definitions:
...
  properties.CommercialData:
    properties:
      ceiling_height:
        type: number
      internal_offices:
        type: integer
      land_use:
        type: string
      loading_docks:
        type: integer
      three_phase_power:
        type: boolean
    type: object
  properties.CreateCommercialInput:
    properties:
      ceiling_height:
        example: 4.5
        type: number
      internal_offices:
        example: 2
        type: integer
      land_use:
        example: Retail
        type: string
      loading_docks:
        example: 1
        type: integer
      three_phase_power:
        example: true
        type: boolean
    type: object
  properties.CreateLocationInput:
    properties:
      city_id:
        example: 1
        type: integer
      exterior_number:
        example: "45"
        type: string
      interior_number:
        example: A
        type: string
      is_public_address:
        example: true
        type: boolean
      latitude:
        example: 19.5438
        type: number
      longitude:
        example: -96.9102
        type: number
      neighborhood:
        example: Centro
        type: string
      postal_code:
        example: "91000"
        type: string
      street:
        example: Av. Principal
        type: string
    type: object
  properties.CreatePropertyClauseInput:
    properties:
      boolean_value:
        example: true
        type: boolean
      clause_id:
        example: 1
        type: integer
      integer_value:
        example: 2
        type: integer
      max_value:
        example: 3
        type: number
      min_value:
        example: 1
        type: number
    type: object
  properties.CreatePropertyInput:
    properties:
      clauses:
        items:
          $ref: '#/definitions/properties.CreatePropertyClauseInput'
        type: array
      commercial:
        $ref: '#/definitions/properties.CreateCommercialInput'
      description:
        example: Spacious residential property near downtown
        type: string
      is_featured:
        example: false
        type: boolean
      location:
        $ref: '#/definitions/properties.CreateLocationInput'
      lot_area:
        example: 200
        type: number
      modality_id:
        example: 1
        type: integer
      property_type_id:
        example: 1
        type: integer
      rent_prices:
        items:
          $ref: '#/definitions/properties.CreateRentPriceInput'
        type: array
      residential:
        $ref: '#/definitions/properties.CreateResidentialInput'
      sale_price:
        $ref: '#/definitions/properties.CreateSalePriceInput'
      services:
        example:
        - 1
        - 3
        - 7
        items:
          type: integer
        type: array
      subtype:
        example: residential
        type: string
      title:
        example: Casa en Xalapa
        type: string
    type: object
  properties.CreatePropertyResult:
    properties:
      data:
        $ref: '#/definitions/properties.CreatePropertyResultData'
    type: object
  properties.CreatePropertyResultData:
    properties:
      property_uuid:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
  properties.CreateRentPriceInput:
    properties:
      currency:
        example: MXN
        type: string
      deposit:
        example: 16000
        type: number
      is_negotiable:
        example: false
        type: boolean
      period_id:
        example: 3
        type: integer
      rent_price:
        example: 8000
        type: number
    type: object
  properties.CreateResidentialInput:
    properties:
      bathrooms:
        example: 2
        type: integer
      bedrooms:
        example: 3
        type: integer
      beds:
        example: 4
        type: integer
      built_area:
        example: 120
        type: number
      construction_year:
        example: 2010
        type: integer
      floors:
        example: 2
        type: integer
      is_furnished:
        example: false
        type: boolean
      orientation_id:
        example: 2
        type: integer
      parking_spots:
        example: 1
        type: integer
    type: object
  properties.CreateSalePriceInput:
    properties:
      currency:
        example: MXN
        type: string
      is_negotiable:
        example: true
        type: boolean
      sale_price:
        example: 1500000
        type: number
    type: object
  properties.DeletePropertyInput:
    properties:
      confirm:
        example: true
        type: boolean
    type: object
...
  properties.GetPropertyData:
    properties:
      commercial:
        $ref: '#/definitions/properties.CommercialData'
      description:
        type: string
      is_featured:
        type: boolean
      location:
        $ref: '#/definitions/properties.LocationData'
      lot_area:
        type: number
      modality_id:
        type: integer
      property_type_id:
        type: integer
      property_uuid:
        type: string
      registered_by:
        type: string
      residential:
        $ref: '#/definitions/properties.ResidentialData'
      status_id:
        type: integer
      subtype:
        type: string
      title:
        type: string
    type: object
  properties.GetPropertyHistoryResult:
    properties:
      data:
        items:
          $ref: '#/definitions/properties.PropertyStatusHistoryData'
        type: array
    type: object
...
  properties.GetPropertyPricesHistoryResult:
    properties:
      data:
        items:
          $ref: '#/definitions/properties.PropertyPriceHistoryData'
        type: array
    type: object
...
  properties.GetPropertyResult:
    properties:
      data:
        $ref: '#/definitions/properties.GetPropertyData'
    type: object
...
  properties.ListPropertiesMeta:
    properties:
      current_page:
        type: integer
      has_next:
        type: boolean
      has_prev:
        type: boolean
      page_size:
        type: integer
      total_count:
        type: integer
      total_pages:
        type: integer
    type: object
  properties.ListPropertiesResult:
    properties:
      data:
        items:
          $ref: '#/definitions/properties.PropertyCardData'
        type: array
      meta:
        $ref: '#/definitions/properties.ListPropertiesMeta'
    type: object
  properties.LocationData:
    properties:
      city_id:
        type: integer
      exterior_number:
        type: string
      interior_number:
        type: string
      is_public_address:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      neighborhood:
        type: string
      postal_code:
        type: string
      street:
        type: string
    type: object
  properties.PropertyCardData:
    properties:
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      built_area:
        type: number
      cover_photo_url:
        example: https://cdn.example.com/properties/cover.jpg
        type: string
      modality:
        $ref: '#/definitions/properties.PropertyCardModalityData'
      price:
        $ref: '#/definitions/properties.PropertyCardPriceData'
      property_type:
        $ref: '#/definitions/properties.PropertyCardTypeData'
      property_uuid:
        example: 123e4567-e89b-12d3-a456-426614174000
        type: string
      status:
        $ref: '#/definitions/properties.PropertyCardStatusData'
      title:
        example: Apartment in Downtown
        type: string
    type: object
  properties.PropertyCardModalityData:
    properties:
      modality_id:
        example: 2
        type: integer
      name:
        example: Rent
        type: string
    type: object
  properties.PropertyCardPriceData:
    properties:
      amount:
        example: 25000
        type: number
      currency:
        example: MXN
        type: string
      period_name:
        example: Monthly
        type: string
      price_type:
        example: rent
        type: string
    type: object
  properties.PropertyCardStatusData:
    properties:
      name:
        example: Available
        type: string
      status_id:
        example: 2
        type: integer
    type: object
  properties.PropertyCardTypeData:
    properties:
      icon:
        example: /icons/apartment.svg
        type: string
      name:
        example: Apartment
        type: string
      property_type_id:
        example: 1
        type: integer
    type: object
...
  properties.PropertyPriceHistoryData:
    properties:
      amount:
        example: 1500000
        type: number
      currency:
        example: MXN
        type: string
      deposit:
        example: 50000
        type: number
      is_current:
        example: true
        type: boolean
      is_negotiable:
        example: true
        type: boolean
      period_name:
        example: Monthly
        type: string
      price_type:
        example: sale
        type: string
      valid_from:
        example: "2026-05-02T10:00:00Z"
        type: string
      valid_until:
        example: "2026-06-02T10:00:00Z"
        type: string
    type: object
  properties.PropertyStatusHistoryData:
    properties:
      changed_at:
        type: string
      changed_by_name:
        type: string
      history_id:
        type: integer
      new_status_name:
        type: string
      previous_status_name:
        type: string
      property_uuid:
        type: string
    type: object
  properties.ResidentialData:
    properties:
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      beds:
        type: integer
      built_area:
        type: number
      construction_year:
        type: integer
      floors:
        type: integer
      is_furnished:
        type: boolean
      orientation_id:
        type: integer
      parking_spots:
        type: integer
    type: object
  properties.UpdateCommercialInput:
    properties:
      ceiling_height:
        type: number
      internal_offices:
        type: integer
      land_use:
        type: string
      loading_docks:
        type: integer
      three_phase_power:
        type: boolean
    type: object
  properties.UpdateLocationInput:
    properties:
      city_id:
        type: integer
      exterior_number:
        type: string
      interior_number:
        type: string
      is_public_address:
        type: boolean
      latitude:
        type: number
      longitude:
        type: number
      neighborhood:
        type: string
      postal_code:
        type: string
      street:
        type: string
    type: object
...
  properties.UpdatePropertyInput:
    properties:
      commercial:
        $ref: '#/definitions/properties.UpdateCommercialInput'
      description:
        type: string
      is_featured:
        type: boolean
      location:
        $ref: '#/definitions/properties.UpdateLocationInput'
      lot_area:
        type: number
      residential:
        $ref: '#/definitions/properties.UpdateResidentialInput'
      title:
        type: string
    type: object
...
  properties.UpdatePropertyResult:
    properties:
      message:
        type: string
    type: object
...
  properties.UpdateResidentialInput:
    properties:
      bathrooms:
        type: integer
      bedrooms:
        type: integer
      beds:
        type: integer
      built_area:
        type: number
      construction_year:
        type: integer
      floors:
        type: integer
      is_furnished:
        type: boolean
      orientation_id:
        type: integer
      parking_spots:
        type: integer
    type: object
...
  shared.ErrorResponse:
    properties:
      error:
        example: error message
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API de Spazio Backend
  title: Spazio API
  version: "1.0"
paths:
...
  /api/v1/properties:
    get:
      description: Returns a paginated list of property cards. Administrators can
        view non-deleted properties, while agents can only view properties assigned
        in property_agents. Supports filtering by search query, status, property type,
        modality, location (country, state, city), price range, and minimum bedrooms.
        Price selection logic prioritizes sale price, then the best current rent price
        (Monthly > Annual > Weekly > Daily).
      parameters:
      - default: 1
        description: Page number (starts at 1)
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: page_size
        type: integer
      - description: Search term (matches title, address, city, state, country)
        in: query
        name: q
        type: string
      - collectionFormat: csv
        description: Filter by status IDs (defaults to Available=2 when omitted)
        in: query
        items:
          type: integer
        name: status_id
        type: array
      - description: Filter by property type ID
        in: query
        name: property_type_id
        type: integer
      - description: Filter by modality ID
        in: query
        name: modality_id
        type: integer
      - description: Filter by country ID
        in: query
        name: country_id
        type: integer
      - description: Filter by state ID
        in: query
        name: state_id
        type: integer
      - description: Filter by city ID
        in: query
        name: city_id
        type: integer
      - description: Minimum price filter
        in: query
        name: min_price
        type: number
      - description: Maximum price filter
        in: query
        name: max_price
        type: number
      - description: Minimum bedrooms filter (Residential only)
        in: query
        name: min_bedrooms
        type: integer
      - description: 'Sort by: created_at, title, price'
        in: query
        name: sort
        type: string
      - description: 'Sort order: asc, desc'
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Paginated list of property cards
          schema:
            $ref: '#/definitions/properties.ListPropertiesResult'
        "400":
          description: Invalid input parameters
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "401":
          description: Missing or invalid authenticated session
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List properties with advanced filters
      tags:
      - Properties
    post:
      consumes:
      - application/json
      description: Registers a property and all related records in a single database
        transaction. The backend generates the property UUID and stores subtype, location,
        pricing, services, and clauses atomically. The authenticated user is set as
        the owner.
      parameters:
      - description: Property payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.CreatePropertyInput'
      produces:
      - application/json
      responses:
        "201":
          description: Property created
          schema:
            $ref: '#/definitions/properties.CreatePropertyResult'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Register a new property
      tags:
      - Properties
  /api/v1/properties/{uuid}:
    delete:
      consumes:
      - application/json
      description: Soft deletes a property by UUID. Requires confirm=true. The authenticated
        admin is recorded as the author of the status change. The operation deletes
        linked photo objects from storage before applying database updates. Only properties
        with available status can be deleted.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Delete payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.DeletePropertyInput'
      produces:
      - application/json
      responses:
        "200":
          description: Property deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete property
      tags:
      - Properties
    get:
      description: Returns property base data, subtype, and location for the given
        UUID. Administrators can view all fields including registered_by. Agents can
        only access assigned properties and do not receive registered_by.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property data
          schema:
            $ref: '#/definitions/properties.GetPropertyResult'
        "400":
          description: Invalid path parameter
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "401":
          description: Missing or invalid authenticated session
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get property by UUID
      tags:
      - Properties
    patch:
      consumes:
      - application/json
      description: |-
        Updates property base data, subtype and location.
        Requires an authenticated admin session.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Update payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/properties.UpdatePropertyInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/properties.UpdatePropertyResult'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update property
      tags:
      - Properties
...
  /api/v1/properties/{uuid}/history:
    get:
      description: Returns the chronological history of status changes for a specific
        property. Requires an authenticated admin session.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Chronological status history retrieved successfully
          schema:
            $ref: '#/definitions/properties.GetPropertyHistoryResult'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "401":
          description: Missing or invalid authenticated session
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get property status history
      tags:
      - Properties
...
  /api/v1/properties/{uuid}/prices/history:
    get:
      description: Returns the complete price history for a property. Requires admin
        role.
      parameters:
      - description: Property UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Property prices history
          schema:
            $ref: '#/definitions/properties.GetPropertyPricesHistoryResult'
        "400":
          description: Invalid path parameter
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "401":
          description: Missing or invalid authenticated session
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "404":
          description: Property not found
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
        "500":
          description: Internal error
          schema:
            $ref: '#/definitions/shared.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get property prices history
      tags:
      - Properties
...
swagger: "2.0"